Переделал все индексы. Добавил макет окна "Настройки подключения к базе данных". Переделал структуру файлов. Переделал страницу с ошибками. И ещё мелкие правки

+++ OrderHandler.DB/Context.cs
@@ -46,7 +46,7 @@ public sealed class Context : DbContext
         modelBuilder.ApplyConfiguration(new CaseNameConfiguration());
     }
 
-    public int GetLastIndex()
+    public long GetLastIndex()
     {
         var lastOrder = Orders.OrderBy(order => order.Id).LastOrDefault();
