+++ OrderHandler.DB/Data/OrderAdd/Milling.cs
@@ -11,7 +11,7 @@ public class Milling : OrderGeneric {
     /// <summary>
     /// Id в БД. Ключ записи
     /// </summary>
-    public override int Id { get; set; }
+    public override long Id { get; set; }
     #region [ Basic Information ]
     /// <summary>
     /// Плановая дата
@@ -31,7 +31,7 @@ public class Milling : OrderGeneric {
     /// Id Пользователя.
     /// Внешний ключ
     /// </summary>
-    public override int UserId { get; set; }
+    public override long UserId { get; set; }
     /// <summary>
     /// Выполнивший пользователь.
     /// Внешний ключ
@@ -41,7 +41,7 @@ public class Milling : OrderGeneric {
     /// Id Заказа.
     /// Внешний ключ
     /// </summary>
-    public override int OrderId { get; set; }
+    public override long OrderId { get; set; }
     /// <summary>
     /// Заказ.
     /// Внешний ключ
