+++ OrderHandler.UI/Model/ViewOrderAdd/ViewAssembling.cs
@@ -15,7 +15,7 @@ public class ViewAssembling : PropertyChanger, IDataErrorInfo {
     public string Error => throw new NotImplementedException();
     IViewAssemblingValidator Validator { get; }
 
-    internal int Id { get; set; }
+    internal long Id { get; set; }
 
     public DateTime PlannedDate {
         get => _plannedDate;
@@ -24,7 +24,7 @@ public class ViewAssembling : PropertyChanger, IDataErrorInfo {
             OnPropertyChanged();
         }
     }
-    internal int? IdUser { get; set; }
+    internal long? IdUser { get; set; }
 
     public string FIO {
         get => _FIO;
