+++ OrderHandler.UI/Model/ViewOrderAdd/ViewEquipment.cs
@@ -14,7 +14,7 @@ public class ViewEquipment : PropertyChanger, IDataErrorInfo {
     public string Error => throw new NotImplementedException();
     IViewEquipmentValidator Validator { get; }
 
-    internal int Id { get; set; }
+    internal long Id { get; set; }
 
     public DateTime PlannedDate {
         get => _plannedDate;
@@ -23,7 +23,7 @@ public class ViewEquipment : PropertyChanger, IDataErrorInfo {
             OnPropertyChanged();
         }
     }
-    internal int? IdUser { get; set; }
+    internal long? IdUser { get; set; }
 
     public string FIO {
         get => _FIO;
