+++ OrderHandler.UI/Model/ViewOrderAdd/ViewPacking.cs
@@ -15,7 +15,7 @@ public class ViewPacking : PropertyChanger, IDataErrorInfo {
     public string Error => throw new NotImplementedException();
     IViewPackingValidator Validator { get; }
 
-    internal int Id { get; set; }
+    internal long Id { get; set; }
 
     public DateTime PlannedDate {
         get => _plannedDate;
@@ -24,7 +24,7 @@ public class ViewPacking : PropertyChanger, IDataErrorInfo {
             OnPropertyChanged();
         }
     }
-    internal int? IdUser { get; set; }
+    internal long? IdUser { get; set; }
 
     public string FIO {
         get => _FIO;
